sensorsmachine.py
import streamlit as st
import pandas as pd
import faiss
import torch
from sentence_transformers import SentenceTransformer
from transformers import pipeline
from reportlab.lib.pagesizes import letter
from reportlab.pdfgen import canvas
import datetime
import random

# ---------- Title and Sidebar ----------
st.title("🛠️ CNC Predictive Maintenance - Multi-Agent System")

st.sidebar.title("📊 Dataset Overview")

# Load Datasets
sensor_data = pd.read_csv("sensor_data.csv")
maintenance_data = pd.read_csv("maintenance_logs.csv")
failure_data = pd.read_csv("failure_records.csv")

# Standardize column names
sensor_data.columns = sensor_data.columns.str.lower()
maintenance_data.columns = maintenance_data.columns.str.lower()
failure_data.columns = failure_data.columns.str.lower()

# Sidebar metrics
st.sidebar.metric("Temperature (°C)", "42")
st.sidebar.metric("Humidity (%)", "63")
st.sidebar.metric("Vibration (g)", "5.2")
st.sidebar.metric("Frequency (Hz)", "120")

# Dataset Shapes
st.sidebar.markdown(f"**Sensor Data:** {sensor_data.shape}")
st.sidebar.markdown(f"**Maintenance Data:** {maintenance_data.shape}")
st.sidebar.markdown(f"**Failure Data:** {failure_data.shape}")

# Dataset Previews
with st.expander("📍 Sensor Data"):
    st.dataframe(sensor_data.head())
with st.expander("🛠️ Maintenance Data"):
    st.dataframe(maintenance_data.head())
with st.expander("⚠️ Failure Data"):
    st.dataframe(failure_data.head())

# ---------- RAG Setup ----------
docs = [
    "CNC machines require routine maintenance to prevent breakdowns.",
    "Vibration sensors help detect anomalies in machine performance.",
    "Humidity affects machine accuracy and lifespan.",
    "Scheduled maintenance reduces downtime and improves efficiency.",
    "Predictive maintenance relies on real-time sensor data.",
]

embedding_model = SentenceTransformer("all-MiniLM-L6-v2")
embedding_model.to(torch.device("cpu"))
doc_embeddings = embedding_model.encode(docs)
index = faiss.IndexFlatL2(doc_embeddings[0].shape[0])
index.add(doc_embeddings)
rag_model = pipeline("text2text-generation", model="google/flan-t5-small", framework="pt")

# ---------- Query Input and Response ----------
st.markdown("### 🔍 Ask the Maintenance System Anything")
query = st.text_input("Type your query below...")
query_button = st.button("Get Response")

if query_button and query:
    query_embedding = embedding_model.encode([query])
    D, I = index.search(query_embedding, k=2)
    retrieved_docs = [docs[i] for i in I[0]]
    context = " ".join(retrieved_docs)
    prompt = f"Context: {context} \n\nQuestion: {query} \nAnswer:"
    response = rag_model(prompt, max_length=100, do_sample=False)[0]["generated_text"]

    st.markdown("### 📖 Retrieved Context")
    st.write(context)
    st.markdown("### 🤖 Answer")
    st.write(response)

# ---------- Anomaly Alerts ----------
def detect_anomalies(sensor_data):
    if sensor_data['vibration'] > 5.0:
        return True, f"Vibration spike detected at {sensor_data['vibration']} g, potential misalignment"
    return False, ""

example_sensor = {'vibration': 5.2, 'temperature': 50}
alert, alert_msg = detect_anomalies(example_sensor)

if alert:
    st.warning(f"🚨 Alert: {alert_msg}")
    st.text("Recommended Action: Schedule bearing inspection within 24 hours.")

# ---------- Maintenance Schedule PDF ----------
def generate_maintenance_schedule(machine_id, task, date, downtime_hours):
    pdf_filename = f"maintenance_schedule_{machine_id}.pdf"
    c = canvas.Canvas(pdf_filename, pagesize=letter)
    c.drawString(100, 750, f"Maintenance Schedule for Machine {machine_id}")
    c.drawString(100, 730, f"Task: {task}")
    c.drawString(100, 710, f"Date: {date}")
    c.drawString(100, 690, f"Estimated Downtime: {downtime_hours} hours")
    c.save()
    return pdf_filename

if st.button("Generate Maintenance Schedule"):
    schedule_pdf = generate_maintenance_schedule(45, "Bearing replacement", "April 28, 2025", 2)
    st.download_button("Download Maintenance Schedule", schedule_pdf)

# ---------- Monthly Report ----------
def generate_performance_report(downtime_reduction, cost_savings, efficiency_gain):
    return f"Monthly Performance Report:\n- Downtime Reduction: {downtime_reduction}%\n- Cost Savings: ${cost_savings}\n- Efficiency Gain: {efficiency_gain}%"

report = generate_performance_report(15, 10000, 10)
st.markdown("### 📊 Monthly Performance Report")
st.write(report)

# ---------- Agent System Status ----------
st.markdown("### 👷 Multi-Agent Pipeline Status")
st.success("✅ All Agents Completed Successfully!")
st.markdown("""
- **Sensor Data Agent:** Successfully read and preprocessed sensor data.
- **Anomaly Detection Agent:** Applied LSTMs and Autoencoders to detect anomalies.
- **Maintenance Scheduling Agent:** Optimized and scheduled maintenance tasks.
- **Alert Notification Agent:** Sent real-time alerts and recommendations to the technician.
""")

# ---------- Footer ----------
st.caption("🔧 Built for Predictive Maintenance of CNC Machines using a Multi-Agent AI System")