requirements.txt
streamlit==1.33.0
pandas==2.2.2
sentence-transformers==2.2.2
transformers==4.39.3
torch==2.2.2
faiss-cpu==1.7.4
reportlab==4.1.0
